backend/requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
gtts
requests
python-multipart
pandas
sentence-transformers
openai
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]